+++ app/Http/Requests/CommodityUpdateRequest.php
@@ -4,9 +4,8 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class EquipmentTypeRequest extends FormRequest
+class CommodityUpdateRequest extends FormRequest
 {
-
     /**
      * Indicates if the validator should stop on the first rule failure.
      *
@@ -32,29 +31,29 @@ class EquipmentTypeRequest extends FormRequest
     public function rules()
     {
         return [
+            'commodity_id' => 'required|profanity|max:5|min:2',
             'status' => 'required',
-            'description' => 'required|max:255',
+            'description' => 'required|profanity|max:255',
         ];
     }
 
     public function messages()
-        {
-            return [
-                'status.required' => 'Status is required',
-                'equip_type_id.required' => 'Equipment Type ID is required',
-                'equip_type_id.unique' => 'Equipment Type ID is already taken',
-                'description.required' => 'Description is required',
-                'description.max' => 'Description must be less than 255 characters',
-            ];
-        }
+    {
+        return [
+            'status.required' => 'Status is required',
+            'commodity_id.required' => 'Commodity ID is required',
+            'description.required' => 'Description is required',
+            'description.max' => 'Description must be less than 255 characters',
+        ];
+    }
 
     public function attributes()
     {
         return [
             'status' => 'Status',
-            'equip_type_id' => 'Equipment Type ID',
+            'commodity_id' => 'Commodity ID',
             'description' => 'Description',
         ];
     }
-}
 
+}