+++ resources/views/commodities/_partials/_content.blade.php
@@ -3,82 +3,14 @@
         <div class="d-flex flex-stack mb-5">
             <!--begin::Search-->
             <div class="d-flex align-items-center position-relative my-1">
-				<span class="svg-icon svg-icon-1 position-absolute ms-6">
-					<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-						<rect opacity="0.5" x="17.0365" y="15.1223" width="8.15546" height="2" rx="1" transform="rotate(45 17.0365 15.1223)" fill="black" />
-						<path d="M11 19C6.55556 19 3 15.4444 3 11C3 6.55556 6.55556 3 11 3C15.4444 3 19 6.55556 19 11C19 15.4444 15.4444 19 11 19ZM11 5C7.53333 5 5 7.53333 5 11C5 14.4667 7.53333 17 11 17C14.4667 17 17 14.4667 17 11C17 7.53333 14.4667 5 11 5Z" fill="black" />
-					</svg>
-				</span>
-                <input type="text" data-kt-docs-table-filter="search" class="form-control form-control-solid w-250px ps-15" placeholder="Search Commodities"/>
             </div>
             <!--end::Search-->
 
             <!--begin::Toolbar-->
             <div class="d-flex justify-content-end" data-kt-docs-table-toolbar="base">
-                <!--begin::Filter-->
-                <button type="button" class="btn btn-light-info me-3 btn-sm" data-kt-menu-trigger="click" data-kt-menu-placement="bottom-end">
-                    <!--begin::Svg Icon | path: icons/duotune/general/gen031.svg-->
-                    <span class="svg-icon svg-icon-3">
-														<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-															<path d="M19.0759 3H4.72777C3.95892 3 3.47768 3.83148 3.86067 4.49814L8.56967 12.6949C9.17923 13.7559 9.5 14.9582 9.5 16.1819V19.5072C9.5 20.2189 10.2223 20.7028 10.8805 20.432L13.8805 19.1977C14.2553 19.0435 14.5 18.6783 14.5 18.273V13.8372C14.5 12.8089 14.8171 11.8056 15.408 10.964L19.8943 4.57465C20.3596 3.912 19.8856 3 19.0759 3Z" fill="black" />
-														</svg>
-													</span>
-                    <!--end::Svg Icon-->Filter</button>
-                <!--begin::Menu 1-->
-                <div class="menu menu-sub menu-sub-dropdown w-300px w-md-325px" data-kt-menu="true" id="kt-toolbar-filter">
-                    <!--begin::Header-->
-                    <div class="px-7 py-5">
-                        <div class="fs-4 text-dark fw-bolder">Filter Options</div>
-                    </div>
-                    <!--end::Header-->
-                    <!--begin::Separator-->
-                    <div class="separator border-gray-200"></div>
-                    <!--end::Separator-->
-                    <!--begin::Content-->
-                    <div class="px-7 py-5">
-                        <!--begin::Input group-->
-                        <div class="mb-10">
-                            <!--begin::Label-->
-                            <label class="form-label fs-5 fw-bold mb-3">Status:</label>
-                            <!--end::Label-->
-                            <!--begin::Options-->
-                            <div class="d-flex flex-column flex-wrap fw-bold" data-kt-docs-table-filter="status">
-                                <!--begin::Option-->
-                                <label class="form-check form-check-sm form-check-custom form-check-solid mb-3 me-5">
-                                    <input class="form-check-input" type="radio" name="status" value="all" checked="checked" />
-                                    <span class="form-check-label text-gray-600">All</span>
-                                </label>
-                                <!--end::Option-->
-                                <!--begin::Option-->
-                                <label class="form-check form-check-sm form-check-custom form-check-solid mb-3 me-5">
-                                    <input class="form-check-input" type="radio" name="status" value="true" />
-                                    <span class="form-check-label text-gray-600">Active</span>
-                                </label>
-                                <!--end::Option-->
-                                <!--begin::Option-->
-                                <label class="form-check form-check-sm form-check-custom form-check-solid mb-3">
-                                    <input class="form-check-input" type="radio" name="status" value="false" />
-                                    <span class="form-check-label text-gray-600">In-active</span>
-                                </label>
-                                <!--end::Option-->
-                            </div>
-                            <!--end::Options-->
-                        </div>
-                        <!--end::Input group-->
-                        <!--begin::Actions-->
-                        <div class="d-flex justify-content-end">
-                            <button type="reset" class="btn btn-light btn-active-light-primary me-2" data-kt-menu-dismiss="true" data-kt-docs-table-filter="reset">Reset</button>
-                            <button type="submit" class="btn btn-primary" data-kt-menu-dismiss="true" data-kt-docs-table-filter="filter">Apply</button>
-                        </div>
-                        <!--end::Actions-->
-                    </div>
-                    <!--end::Content-->
-                </div>
-                <!--end::Menu 1-->
-                <!--end::Filter-->
                 <!--begin::Add customer-->
                 @can('commodities-create')
-                    <button type="button" class="btn btn-light-primary btn-sm" data-bs-toggle="modal" data-bs-target="#create-commodity">
+                    <button type="button" class="btn btn-light-primary btn-sm me-3" data-bs-toggle="modal" data-bs-target="#create-commodity">
                         <!--begin::Svg Icon | path: icons/duotune/arrows/arr075.svg-->
                         <span class="svg-icon svg-icon-3">
 														<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
@@ -89,16 +21,49 @@
                         <!--end::Svg Icon-->Add Commodity</button>
             @endcan
             <!--end::Add customer-->
+                <button type="button" class="btn btn-icon btn-light-info btn-sm"
+                        data-kt-menu-trigger="click"
+                        data-kt-menu-placement="bottom-start">
+                    <!--begin::Svg Icon | path: assets/media/icons/duotune/communication/com008.svg-->
+                    <span class="svg-icon svg-icon-1">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+                            <rect opacity="0.3" x="2" y="2" width="20" height="20" rx="10" fill="black"/>
+                            <rect x="11" y="11" width="2" height="2" rx="1" fill="black"/>
+                            <rect x="11" y="15" width="2" height="2" rx="1" fill="black"/>
+                            <rect x="11" y="7" width="2" height="2" rx="1" fill="black"/>
+                        </svg>
+                    </span>
+                    <!--end::Svg Icon-->
+                </button>
+                <!--end::Trigger-->
+                <!--begin::Menu-->
+                <div class="menu menu-sub menu-sub-dropdown menu-column menu-rounded menu-gray-600 menu-state-bg-light-primary fw-bold fs-7 w-200px py-4"
+                     data-kt-menu="true">
+                    <!--begin::Menu item-->
+                    <div class="menu-item px-3">
+                        <a data-bs-dismiss="modal" data-bs-toggle="tooltip" data-bs-placement="left" title="Coming Soon..." class="menu-link px-3">
+                            Import From Excel
+                        </a>
+                    </div>
+                    <!--end::Menu item-->
+                    <!--begin::Menu item-->
+                    <div class="menu-item px-3">
+                        <a href="#" class="menu-link px-3">
+                            Export To Excel
+                        </a>
+                    </div>
+                    <!--end::Menu item-->
+                </div>
+                <!--end::Menu-->
             </div>
             <!--end::Toolbar-->
             <!--end::Group actions-->
         </div>
         <!--end::Wrapper-->
         <div class="table-responsive">
-            <table id="commodities_table" class="table align-middle table-row-dashed fs-6 gy-5">
+            <table id="commodity-table" class="table align-middle table-row-dashed fs-6 gy-5">
                 <thead>
                 <tr class="fw-bold fs-6 text-gray-800 border-bottom border-gray-200">
-                    <th>No</th>
                     <th>Status</th>
                     <th>ID</th>
                     <th>Description</th>
@@ -106,37 +71,7 @@
                 </tr>
                 </thead>
                 <tbody class="text-gray-600 fw-bold">
-                @foreach ($commodities as $commodity)
-                    <tr>
-                        <td>{{ ++$i }}</td>
-                        <td>{{ $commodity->status }}</td>
-                        <td>{{ $commodity->commodity_id }}</td>
-                        <td>{{ $commodity->description }}</td>
-                        <!--begin::Action=-->
-                        <td class="text-end">
-                            <a href="#" class="btn btn-light btn-active-light-primary btn-sm" data-kt-menu-trigger="click" data-kt-menu-placement="bottom-end">Actions
-                                <!--begin::Svg Icon | path: icons/duotune/arrows/arr072.svg-->
-                                <span class="svg-icon svg-icon-5 m-0">
-															<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-																<path d="M11.4343 12.7344L7.25 8.55005C6.83579 8.13583 6.16421 8.13584 5.75 8.55005C5.33579 8.96426 5.33579 9.63583 5.75 10.05L11.2929 15.5929C11.6834 15.9835 12.3166 15.9835 12.7071 15.5929L18.25 10.05C18.6642 9.63584 18.6642 8.96426 18.25 8.55005C17.8358 8.13584 17.1642 8.13584 16.75 8.55005L12.5657 12.7344C12.2533 13.0468 11.7467 13.0468 11.4343 12.7344Z" fill="black" />
-															</svg>
-														</span>
-                                <!--end::Svg Icon--></a>
-                            <!--begin::Menu-->
-                            <div class="menu menu-sub menu-sub-dropdown menu-column menu-rounded menu-gray-600 menu-state-bg-light-primary fw-bold fs-7 w-125px py-4" data-kt-menu="true">
-                                <!--begin::Menu item-->
-                                @can('commodities-edit')
-                                    <div class="menu-item px-3">
-                                        <a data-bs-toggle="modal" data-bs-target="#edit-commodity-{{ $commodity->id }}" class="menu-link px-3">Edit</a>
-                                    </div>
-                            @endcan
-                            <!--end::Menu item-->
-                            </div>
-                            <!--end::Menu-->
-                        </td>
-                        <!--end::Action=-->
-                    </tr>
-                @endforeach
+
                 </tbody>
             </table>
         </div>