+++ resources/views/equipmenttype/index.blade.php
@@ -31,3 +31,34 @@ $(function(){$('#equip-type-table').DataTable({processing:!0,serverSide:!0,searc
 </script>
 <!-- end::Script to Produce DataTable for Equipment Types -->
 
+<script type="text/javascript">
+
+    // CSRF Token
+    var CSRF_TOKEN = $('meta[name="csrf-token"]').attr('content');
+    $(document).ready(function(){
+
+        $( "#selEquip" ).select2({
+            ajax: {
+                url: "{{route('equipmenttype.showEquipTypes')}}",
+                type: "post",
+                dataType: 'json',
+                delay: 250,
+                data: function (params) {
+                    return {
+                        _token: CSRF_TOKEN,
+                        search: params.term // search term
+                    };
+                },
+                processResults: function (response) {
+                    return {
+                        results: response
+                    };
+                },
+                cache: true
+            }
+
+        });
+
+    });
+</script>
+